# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/files": {
      "get": {
        "operationId": "FileManagerController_listFiles",
        "parameters": [
          {
            "name": "path",
            "required": false,
            "in": "query",
            "description": "Relative folder path (empty for root)",
            "schema": {
              "example": "images/products",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Directory entries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 200
                    },
                    "data": {
                      "$ref": "#/components/schemas/ListFilesResponseDto"
                    },
                    "message": {
                      "type": "string",
                      "example": "OK"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data",
                    "message"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Missing permission",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 403
                    },
                    "message": {
                      "type": "string",
                      "example": "Forbidden"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "statusCode",
                    "message"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Folder not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Not Found"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "statusCode",
                    "message"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "List files and folders in a directory",
        "tags": [
          "FileManager"
        ]
      }
    },
    "/files/upload": {
      "post": {
        "operationId": "FileManagerController_uploadFiles",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "description": "CSRF token from GET /auth/csrf — copy the csrfToken field into this header.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "path": {
                    "type": "string",
                    "description": "Target folder relative path (empty for root)",
                    "example": "images"
                  },
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Uploaded files",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 201
                    },
                    "data": {
                      "$ref": "#/components/schemas/UploadFilesResponseDto"
                    },
                    "message": {
                      "type": "string",
                      "example": "Created"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data",
                    "message"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Missing permission",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 403
                    },
                    "message": {
                      "type": "string",
                      "example": "Forbidden"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "statusCode",
                    "message"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Upload one or more files",
        "tags": [
          "FileManager"
        ]
      }
    },
    "/files/folders": {
      "post": {
        "operationId": "FileManagerController_createFolder",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "description": "CSRF token from GET /auth/csrf — copy the csrfToken field into this header.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFolderRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Folder created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 201
                    },
                    "data": {
                      "$ref": "#/components/schemas/CreateFolderResponseDto"
                    },
                    "message": {
                      "type": "string",
                      "example": "Created"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data",
                    "message"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Missing permission",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 403
                    },
                    "message": {
                      "type": "string",
                      "example": "Forbidden"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "statusCode",
                    "message"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Create a folder",
        "tags": [
          "FileManager"
        ]
      },
      "delete": {
        "operationId": "FileManagerController_deleteFolder",
        "parameters": [
          {
            "name": "path",
            "required": true,
            "in": "query",
            "description": "Relative folder path to delete",
            "schema": {
              "example": "images/banners",
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "description": "CSRF token from GET /auth/csrf — copy the csrfToken field into this header.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Folder deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 200
                    },
                    "data": {
                      "$ref": "#/components/schemas/DeleteFolderResponseDto"
                    },
                    "message": {
                      "type": "string",
                      "example": "OK"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data",
                    "message"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Missing permission",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 403
                    },
                    "message": {
                      "type": "string",
                      "example": "Forbidden"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "statusCode",
                    "message"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Delete a folder and its contents",
        "tags": [
          "FileManager"
        ]
      }
    },
    "/files/{id}": {
      "delete": {
        "operationId": "FileManagerController_deleteFile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "description": "CSRF token from GET /auth/csrf — copy the csrfToken field into this header.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 200
                    },
                    "data": {
                      "$ref": "#/components/schemas/DeleteFileResponseDto"
                    },
                    "message": {
                      "type": "string",
                      "example": "OK"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data",
                    "message"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Missing permission",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 403
                    },
                    "message": {
                      "type": "string",
                      "example": "Forbidden"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "statusCode",
                    "message"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "File not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Not Found"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "statusCode",
                    "message"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Delete a file by id",
        "tags": [
          "FileManager"
        ]
      }
    },
...
  },
  "info": {
    "title": "Lio API",
    "description": "Authentication and session management API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "schemas": {
...
      "EntryDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "folder",
              "file"
            ],
            "example": "file"
          },
          "name": {
            "type": "string",
            "example": "photo.png"
          },
          "path": {
            "type": "string",
            "example": "images/photo-abc123.png"
          },
          "id": {
            "type": "number",
            "example": 1
          },
          "size": {
            "type": "number",
            "example": 2048
          },
          "mimeType": {
            "type": "string",
            "example": "image/png"
          },
          "url": {
            "type": "string",
            "example": "/uploads/images/photo-abc123.png"
          },
          "createdAt": {
            "type": "string",
            "example": "2026-09-02T12:00:00.000Z"
          }
        },
        "required": [
          "type",
          "name",
          "path"
        ]
      },
      "ListFilesResponseDto": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string",
            "example": "images"
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EntryDto"
            }
          }
        },
        "required": [
          "path",
          "entries"
        ]
      },
      "FileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "originalName": {
            "type": "string",
            "example": "photo.png"
          },
          "storedName": {
            "type": "string",
            "example": "a1b2c3d4e5f6a7b8c9d0e1f2a3b4c5d6-9f3a2b1c.png"
          },
          "path": {
            "type": "string",
            "example": "images/a1b2c3d4e5f6a7b8c9d0e1f2a3b4c5d6-9f3a2b1c.png"
          },
          "url": {
            "type": "string",
            "example": "/uploads/images/a1b2c3d4e5f6a7b8c9d0e1f2a3b4c5d6-9f3a2b1c.png"
          },
          "mimeType": {
            "type": "string",
            "example": "image/png"
          },
          "size": {
            "type": "number",
            "example": 2048
          },
          "uploaderId": {
            "type": "object",
            "example": 1,
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "example": "2026-09-02T12:00:00.000Z"
          }
        },
        "required": [
          "id",
          "originalName",
          "storedName",
          "path",
          "url",
          "mimeType",
          "size",
          "createdAt"
        ]
      },
      "UploadFilesResponseDto": {
        "type": "object",
        "properties": {
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileDto"
            }
          }
        },
        "required": [
          "files"
        ]
      },
      "CreateFolderRequestDto": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string",
            "description": "Relative folder path to create",
            "example": "images/banners"
          }
        },
        "required": [
          "path"
        ]
      },
      "CreateFolderResponseDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          },
          "path": {
            "type": "string",
            "example": "images/banners"
          }
        },
        "required": [
          "ok",
          "path"
        ]
      },
      "DeleteFolderResponseDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "ok"
        ]
      },
      "DeleteFileResponseDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "ok"
        ]
      },
...
    }
  }
}